requirements.txt
blinker==1.9.0
certifi==2024.8.30
Flask==3.1.0
requests==2.32.3
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pymongo==4.10.1
urllib3==2.2.3
Werkzeug==3.1.3
beautifulsoup4==4.12.3
selenium==4.27.1
python-dotenv==1.0.1